backend/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
genkit-ai==0.6.2
google-generativeai==0.7.2
pandas==2.2.2
numpy==2.0.0
gunicorn==22.0.0
firebase-admin==6.5.0
pydantic==2.8.2
